fix(): refactoring of session middleware and new access api
